+++ sdk/appconfiguration/app-configuration/samples/helloworldWithLabels.ts
@@ -1,6 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+// This sample builds on concepts in helloworld.ts and shows you how 
+// to use labels. Labels allow you to add an extra dimension for your 
+// setting and gives you a simple way to create conventions for environments. 
+// 
+// For additional information about how labels work see:
+// https://docs.microsoft.com/en-us/azure/azure-app-configuration/concept-key-value#label-keys
+
 // NOTE: replace with import { AppConfigurationClient } from "@azure/app-configuration"
 // in a standalone project
 import { AppConfigurationClient } from "../src"