+++ sdk/appconfiguration/app-configuration/samples/setReadOnlySample.ts
@@ -1,6 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+// This sample shows you how to set a configuration setting to read-only. 
+// This can help prevent accidental deletion or modification of a setting.
+
 // NOTE: replace with import { AppConfigurationClient } from "@azure/app-configuration"
 // in a standalone project
 import { AppConfigurationClient } from "../src";
@@ -13,30 +16,33 @@ export async function run() {
   const client = new AppConfigurationClient(connectionString);
 
   const readOnlySampleKey = "readOnlySampleKey";
-
   await cleanupSampleValues([readOnlySampleKey], client);
 
   console.log("Creating a new key. By default keys are writeable");
   await client.addConfigurationSetting({ key: readOnlySampleKey, label: "a label", value: "Initial value" });
 
   // now we'd like to prevent future modifications - let's set the key/label to read-only
+  console.log("Setting a key to read-only. Any modifications will fail");
   await client.setReadOnly({ key: readOnlySampleKey, label: "a label" });
 
   // any modifications to the key will now throw errors
   try {
+    console.log("Attempting to modify a read-only setting");
     await client.setConfigurationSetting({ key: readOnlySampleKey, label: "a label", value: "new value" });
   } catch (err) {
-    console.log(`Error thrown when trying to modifying a read-only key`, err.name);
+    console.log(`Error gets thrown - can't modify a read-only setting`);
   }
     
   // clients that read from the key are unaffected
   await client.getConfigurationSetting(readOnlySampleKey, { label: "a label" });
-
+  
   // to make a key writable again we can clear the read-only status
+  console.log("Clearing the read-only status on the key so we can update the value");
   await client.clearReadOnly({ key: readOnlySampleKey, label: "a label" });
 
   // and now clients can change the value again
-  await client.setConfigurationSetting({ key: readOnlySampleKey, label: "a label", value: "new value" });
+  const updatedSetting = await client.setConfigurationSetting({ key: readOnlySampleKey, label: "a label", value: "new value" });
+  console.log(`Updated value is ${updatedSetting.value}`);
   
   await cleanupSampleValues([readOnlySampleKey], client);
 }