+++ sdk/identity/identity/src/credentials/authorizationCodeCredential.ts
@@ -34,7 +34,7 @@ export class AuthorizationCodeCredential implements TokenCredential {
    * the authorization code flow to obtain an authorization code to be used
    * with this credential.  A full example of this flow is provided here:
    *
-   * https://github.com/daviwil/azure-sdk-for-js/blob/master/sdk/identity/identity/samples/authorizationCodeSample.ts
+   * https://github.com/Azure/azure-sdk-for-js/blob/master/sdk/identity/identity/samples/authorizationCodeSample.ts
    *
    * @param tenantId The Azure Active Directory tenant (directory) ID or name.
    * @param clientId The client (application) ID of an App Registration in the tenant.