+++ sdk/core/core-arm/Changelog.md
@@ -1,13 +1,17 @@
 # Changelog
 
+## 1.0.0-preview.4 - 2019-10-07
+
+- Updated to use the latest version of the `@azure/core-http` library.
+
 ## 1.0.0-preview.3 - 2019-09-09
 
 - Reintroduces `ServiceClientCredentials` type to `credentials` parameter of
-`AzureServiceClient` ([PR #4773](https://github.com/Azure/azure-sdk-for-js/pull/4773)).
+  `AzureServiceClient` ([PR #4773](https://github.com/Azure/azure-sdk-for-js/pull/4773)).
 
 ## 1.0.0-preview.2 - 2019-08-05
 
-- Removed `ServiceClientCredentials` type from `credentials` parameter of `AzureServiceClient` ([PR #4367](https://github.com/Azure/azure-sdk-for-js/pull/4367)).  Credential implementations are now standardized on `@azure/core-auth`'s `TokenCredential` interface and provided by `@azure/identity`.
+- Removed `ServiceClientCredentials` type from `credentials` parameter of `AzureServiceClient` ([PR #4367](https://github.com/Azure/azure-sdk-for-js/pull/4367)). Credential implementations are now standardized on `@azure/core-auth`'s `TokenCredential` interface and provided by `@azure/identity`.
 
 ## 1.0.0-preview.1 - 2019-07-10